+++ sw/simple-unreal/pid.h
@@ -33,10 +33,10 @@ class PID
         // dt -  loop interval time
         // max - maximum value of manipulated variable
         // min - minimum value of manipulated variable
-        PID( double dt, double max, double min, double Kp, double Kd, double Ki );
+        PID( float dt, float max, float min, float Kp, float Kd, float Ki );
 
         // Returns the manipulated variable given a setpoint and current process value
-        double calculate( double setpoint, double pv );
+        float calculate( float setpoint, float pv );
         ~PID();
 
     private:
